Added user impersonation UI for super admins.
Super admins can impersonate other users for testing purposes. e.g.
to see what they can see in the UI. This impersonation was previously
supported, but only by knowing the correct URL to visit. This commit
adds a drop down menu at the top right when a super user is logged in.
The menu shows a list of users who can be impersonated.